Move BsnVersion to exported package c.y.jdisc.application
