+++ container-core/src/main/java/com/yahoo/container/core/config/ApplicationBundleLoader.java
@@ -3,7 +3,7 @@ package com.yahoo.container.core.config;
 
 import com.yahoo.config.FileReference;
 import com.yahoo.container.di.Osgi.GenerationStatus;
-import com.yahoo.jdisc.core.BsnVersion;
+import com.yahoo.jdisc.application.BsnVersion;
 import com.yahoo.osgi.Osgi;
 import org.osgi.framework.Bundle;
